Add: compress_bed decision; raw chromsome distribution

+++ pipeline/CUT_RUN_TF/CUT_RUN_TF.sh
@@ -24,12 +24,14 @@ function reads_file_process {
 function compress_bed {
     bedFile=${1}
     genomeVersion=${2}
-    col=`head -1 ${bedFile} | awk '{print NF}'`
-    plus=`bc <<< "$col -3"`
-    intersectBed -a ${bedFile} -b <(awk '{print $1"\t0\t"$2}' ~/source/bySpecies/${genomeVersion}/${genomeVersion}.chrom.sizes) -wa -f 1.00 | sort -k1,1 -k2,2n > ${bedFile}.tmp
-    bedToBigBed -type=bed3+${plus} ${bedFile}.tmp ~/source/bySpecies/${genomeVersion}/${genomeVersion}.chrom.sizes ${bedFile::(${#bedFile}-2)}b
-    rm ${bedFile}.tmp
-    # rm ${bedFile}
+    if [[ ! -e ${bedFile::(${#bedFile}-2)}b ]]; then
+        col=`head -1 ${bedFile} | awk '{print NF}'` && \
+        plus=`bc <<< "$col -3"` && \
+        intersectBed -a ${bedFile} -b <(awk '{print $1"\t0\t"$2}' ~/source/bySpecies/${genomeVersion}/${genomeVersion}.chrom.sizes) -wa -f 1.00 | sort -k1,1 -k2,2n > ${bedFile}.tmp && \
+        bedToBigBed -type=bed3+${plus} ${bedFile}.tmp ~/source/bySpecies/${genomeVersion}/${genomeVersion}.chrom.sizes ${bedFile::(${#bedFile}-2)}b && \
+        rm ${bedFile}.tmp # && \
+        # rm ${bedFile}
+    fi
 }
 
 # ----- parameters -----
@@ -87,11 +89,12 @@ function mapping_filtering {
     if [[ ! -e 2_signal/${name}_raw_fragments.bed  ]]; then
         bamToBed -bedpe -i 1_mapping/${name}.bam | awk '$1 !~ /_/{if($2<$5) print $1"\t"$2"\t"$6; else print $1"\t"$5"\t"$3}' > 2_signal/${name}_raw_fragments.bed
     fi
+    cut -f 1 2_signal/${name}_raw_fragments.bed | sort -S 1% | uniq -c | sort -S 1% -k1,1rg | awk 'BEGIN{print "chromosome\tnumber"} {print $2"\t"$1}' > 2_signal/${name}_raw_chromosome_distribution.txt
     if [[ ! -e 2_signal/${name}_fragments.bed  ]]; then
         sort -S 1% -k1,1 -k2,2n 2_signal/${name}_raw_fragments.bed | uniq > 2_signal/${name}_fragments.bed
-        cut -f 1 2_signal/${name}_fragments.bed | sort -S 1% | uniq -c | sort -S 1% -k1,1rg | awk 'BEGIN{print "chromosome\tnumber"} {print $2"\t"$1}' > 2_signal/${name}_chromosome_distribution.txt
-        awk '{print $3-$2}' 2_signal/${name}_fragments.bed | sort -S 1% | uniq -c | sort -S 1% -k2,2g | awk 'BEGIN{print "fragment_length\tnumber"} {print $2"\t"$1}' > 2_signal/${name}_fragments_length.txt
     fi
+    cut -f 1 2_signal/${name}_fragments.bed | sort -S 1% | uniq -c | sort -S 1% -k1,1rg | awk 'BEGIN{print "chromosome\tnumber"} {print $2"\t"$1}' > 2_signal/${name}_chromosome_distribution.txt
+    awk '{print $3-$2}' 2_signal/${name}_fragments.bed | sort -S 1% | uniq -c | sort -S 1% -k2,2g | awk 'BEGIN{print "fragment_length\tnumber"} {print $2"\t"$1}' > 2_signal/${name}_fragments_length.txt
 }
 
 
@@ -114,7 +117,6 @@ function piling_up {
 # ----- short fragments -----
 function short_fragments {
     if [[ ! -e 2_signal/${name}_OCR.bw ]]; then
-        echo B start
         cd 2_signal
         fragment_length=`awk 'BEGIN{s=0;c=0} NR>1{if($1<=120) {s+=$1*$2;c+=$2}} END{printf "%d", s/c}' ${name}_fragments_length.txt`
         awk '{if($3-$2<=120) print}' ${name}_fragments.bed > ${name}_OCR_fragments.bed
@@ -126,7 +128,6 @@ function short_fragments {
         genomeCoverageBed -bga -scale $c -i ${name}_OCR_fragments_shift.bed -g ~/source/bySpecies/${genomeVersion}/${genomeVersion}.chrom.sizes | awk '{if($3>$2) print$0}' > ${name}_OCR_fragments_shift.bdg && \
         ${MY_PATH}/../utilities/bdg2bw.sh ${name}_OCR_fragments_shift.bdg ~/source/bySpecies/${genomeVersion}/${genomeVersion}_main.chrom.sizes ${name}_OCR && \
         rm ${name}_OCR_fragments_shift.bdg ${name}_OCR_fragments_shift.bed
-        echo B end
         cd ..
     fi
 }
@@ -134,6 +135,7 @@ function short_fragments {
 # ----- clearning_up -----
 function clearning_up {
     # rm 1_mapping/${name}.bam
+    compress_bed 2_signal/${name}_raw_fragments.bed ${genomeVersion}
     compress_bed 2_signal/${name}_fragments.bed ${genomeVersion}
     compress_bed 2_signal/${name}_OCR_fragments.bed ${genomeVersion}
 }
